backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
pydantic==2.8.2
pypdf==4.2.0
python-docx==1.1.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
torch
requests==2.32.3
openai==1.35.7